refactor(chain): P2 结构治理 — 拆分 message.py 上帝文件，MediaInteractionChain 独立成模块（审计 #61）
将 MediaInteractionChain（1461 行）从 app/chain/message.py 抽出为新模块
app/chain/media_interaction.py，类体逐字节不变（纯移动）。message.py 由 3369
行降至 1900 行（<2000 行上帝文件阈值），ruff F401 核验删除随类移出后不再使用的
17 个 import；经再导出保持 from app.chain.message import MediaInteractionChain
对外导入路径不变，并供 MessageChain 内部实例化。

tests/test_media_interaction.py：7 处 mock patch 目标从 app.chain.message 重定向
至 app.chain.media_interaction（DirectoryHelper/DownloadChain 协作者随类迁移，
mock 须打在名字被查找的模块）。

验证：全量 1970 passed/19 skipped（与基线一致）；MediaInteractionChain 类体与
MessageChain 体均逐字节不变；ruff 净；3 维度对抗审查 0 新回归 —— 行为保持。

P2 其余项：#60 双轨已于 2cc5ccf6 完成；#16 拆 ChainBase（可行为保持但需重排
~1300 行核心基类、28 子类依赖，diff 大）按分析建议留独立 PR；#9 拆 Scheduler/
Monitor 与 #38 扩展点登记表存在无法行为保持的特例，本轮不做。
