+++ tests/test_media_interaction.py
@@ -289,10 +289,10 @@ def test_torrent_selection_prompts_download_dir_buttons_before_download():
     request.phase = "torrent"
 
     with patch(
-        "app.chain.message.DirectoryHelper.get_download_dirs",
+        "app.chain.media_interaction.DirectoryHelper.get_download_dirs",
         return_value=_build_download_dirs(),
     ), patch.object(chain, "post_message") as post_message, patch(
-        "app.chain.message.DownloadChain.download_single"
+        "app.chain.media_interaction.DownloadChain.download_single"
     ) as download_single:
         handled = chain.handle_text_interaction(
             channel=MessageChannel.Telegram,
@@ -331,7 +331,7 @@ def test_torrent_selection_prompts_text_download_dir_for_plain_channel():
     request.phase = "torrent"
 
     with patch(
-        "app.chain.message.DirectoryHelper.get_download_dirs",
+        "app.chain.media_interaction.DirectoryHelper.get_download_dirs",
         return_value=_build_download_dirs(),
     ), patch.object(chain, "post_message") as post_message:
         handled = chain.handle_text_interaction(
@@ -370,10 +370,10 @@ def test_download_dir_callback_runs_pending_single_download_with_save_path():
     request.pending_download_context = context
 
     with patch(
-        "app.chain.message.DirectoryHelper.get_download_dirs",
+        "app.chain.media_interaction.DirectoryHelper.get_download_dirs",
         return_value=_build_download_dirs(),
     ), patch(
-        "app.chain.message.DownloadChain.download_single",
+        "app.chain.media_interaction.DownloadChain.download_single",
         return_value="hash",
     ) as download_single:
         request.download_dirs = chain._get_download_dirs()
@@ -412,10 +412,10 @@ def test_download_dir_text_reply_runs_pending_single_download_with_save_path():
     request.pending_download_context = context
 
     with patch(
-        "app.chain.message.DirectoryHelper.get_download_dirs",
+        "app.chain.media_interaction.DirectoryHelper.get_download_dirs",
         return_value=_build_download_dirs(),
     ), patch(
-        "app.chain.message.DownloadChain.download_single",
+        "app.chain.media_interaction.DownloadChain.download_single",
         return_value="hash",
     ) as download_single:
         request.download_dirs = chain._get_download_dirs()
